Moved get lable and price fn to util for reuse
