+++ src/components/util/util.js
@@ -14,3 +14,28 @@ export function longDecimalToShort (decimal){
     }
     return decimal;
 }
+
+
+export function getTime(date) {
+    let hour = date.getHours();
+    let munutes = date.getMinutes().toString().padStart(2, "0");
+    let timeString = `${hour} : ${munutes} AM`;
+    if (hour > 12) {
+      timeString = `${(hour - 12).toString().padStart(2, "0")} : ${munutes} PM`;
+    }
+    return timeString;
+  }
+
+  export  function getPriceLabels(arr, days) {
+    const labels = arr.map((item) => {
+      let date = new Date(item[0]);
+      let time = getTime(date);
+      return days == 1 ? time : date.toLocaleDateString();
+    });
+    return labels;
+  }
+
+  export  function getPriceData(arr) {
+    let priceData = arr.map((item) => item[1]);
+    return priceData;
+  }
